fix sync menu

+++ WebApp/WebApp/Controllers/AdminController.cs
@@ -174,35 +174,42 @@ namespace WebApp.Controllers
                     var ids = existingMenu.Select(d => d.CloverId).ToList();
                     ids.RemoveAll(s => s == null);
 
-                    var newMenu = result.Select(r => r.Id).ToList();
-
-                    var itemsToBeDeleted = ids.Where(m => !newMenu.Contains(m)).ToList();
-                    var itemsToBeAdded = newMenu.Where(m => !ids.Contains(m)).ToList();
-
-                    foreach (var i in itemsToBeAdded)
+                    if (result != null)
                     {
-                        var r = result.First(re => re.Id == i);
-                        context.Dishes.Add(new Dish
+                        var newMenu = result.Select(r => r.Id).ToList();
+
+                        var itemsToBeDeleted = ids.Where(m => !newMenu.Contains(m)).ToList();
+                        var itemsToBeAdded = newMenu.Where(m => !ids.Contains(m)).ToList();
+
+                        foreach (var i in itemsToBeAdded)
                         {
-                            Name = r.Name,
-                            Price = r.Price / 100.0,
-                            CuisineId = -1,
-                            RestaurantId = restaurantId,
-                            CloverId = r.Id,
-                            Deleted = false
-                        });
-                    }
+                            var r = result.First(re => re.Id == i);
+                            context.Dishes.Add(new Dish
+                            {
+                                Name = r.Name,
+                                Price = r.Price / 100.0,
+                                CuisineId = -1,
+                                RestaurantId = restaurantId,
+                                CloverId = r.Id,
+                                Deleted = false
+                            });
+                        }
 
-                    foreach (var i in itemsToBeDeleted)
+                        foreach (var i in itemsToBeDeleted)
+                        {
+                            var r = existingMenu.First(re => re.CloverId == i);
+                            context.Dishes.Remove(r);
+                        }
+
+                        context.SaveChanges();
+                    }
+                    else
                     {
-                        var r = existingMenu.First(re => re.CloverId == i);
-                        context.Dishes.Remove(r);
+                        // nothing
                     }
-
-                    context.SaveChanges();
                 }
             }
-            return Ok();
+            return Ok("Success");
         }
     }
 }
